Extract keyboard input main-thread cleanup
Move keyboard state restoration, timeout cleanup, and press overlay finalization into a stateless helper so the use case focuses on orchestration. Track the helper with the async cancellation-token guard.
