+++ Packages/src/Editor/FirstPartyTools/SimulateKeyboard/KeyboardInputMainThreadCleanup.cs
@@ -1,0 +1,163 @@
+#if ULOOP_HAS_INPUT_SYSTEM
+#nullable enable
+using System.Threading;
+using System.Threading.Tasks;
+using UnityEditor;
+using UnityEngine;
+using UnityEngine.InputSystem;
+
+using io.github.hatayama.UnityCliLoop.Runtime;
+using io.github.hatayama.UnityCliLoop.ToolContracts;
+
+namespace io.github.hatayama.UnityCliLoop.FirstPartyTools
+{
+    /// <summary>
+    /// Restores keyboard device and overlay state on the Unity main thread after simulation ends.
+    /// </summary>
+    internal static class KeyboardInputMainThreadCleanup
+    {
+        internal static async Task FinalizePressOverlay(CancellationToken ct)
+        {
+            await InputSystemUpdateHelper.SwitchToMainThreadIfNeeded(CancellationToken.None);
+            if (ct.IsCancellationRequested)
+            {
+                SimulateKeyboardOverlayState.ClearPress();
+                return;
+            }
+
+            SimulateKeyboardOverlayState.ReleasePress();
+            await EditorFrameWaiter.WaitFramesOrTimeoutAsync(
+                1,
+                UnityCliLoopConstants.EDITOR_FRAME_WAIT_TIMEOUT_MS,
+                CancellationToken.None).ConfigureAwait(false);
+        }
+
+        internal static async Task<InputSimulationWaitOutcome> RollbackHeldKey(
+            Keyboard keyboard,
+            Key key,
+            string keyName,
+            CancellationToken ct)
+        {
+            await InputSystemUpdateHelper.SwitchToMainThreadIfNeeded(ct);
+            InputSimulationWaitOutcome releaseOutcome =
+                await ReleaseKeyStateIfPossible(keyboard, key, ct).ConfigureAwait(false);
+            if (releaseOutcome == InputSimulationWaitOutcome.TimedOut)
+            {
+                ScheduleTimedOutHeldKeyCleanup(keyboard, key, keyName, false);
+                return releaseOutcome;
+            }
+
+            await InputSystemUpdateHelper.SwitchToMainThreadIfNeeded(ct);
+            KeyboardKeyState.SetKeyUp(key);
+            SimulateKeyboardOverlayState.RemoveHeldKey(keyName);
+            return releaseOutcome;
+        }
+
+        internal static async Task<InputSimulationWaitOutcome> ReleaseKeyStateIfPossible(
+            Keyboard keyboard,
+            Key key,
+            CancellationToken ct)
+        {
+            await InputSystemUpdateHelper.SwitchToMainThreadIfNeeded(ct);
+            if (!CanInjectKeyboardState(keyboard))
+            {
+                return InputSimulationWaitOutcome.Completed;
+            }
+
+            if (EditorApplication.isPaused)
+            {
+                ReleaseKeyStateImmediately(keyboard, key);
+                return InputSimulationWaitOutcome.Completed;
+            }
+
+            InputSimulationWaitOutcome releaseOutcome = await InputSystemUpdateHelper.ApplyOnNextConfiguredUpdate(
+                () => KeyboardKeyState.SetKeyState(keyboard, key, false),
+                ct).ConfigureAwait(false);
+            if (releaseOutcome == InputSimulationWaitOutcome.TimedOut)
+            {
+                ScheduleReleaseKeyStateImmediately(keyboard, key);
+            }
+
+            return releaseOutcome;
+        }
+
+        private static void ScheduleReleaseKeyStateImmediately(Keyboard keyboard, Key key)
+        {
+            ReleaseKeyStateImmediatelyOnMainThreadAsync(keyboard, key, CancellationToken.None).Forget();
+        }
+
+        private static async Task ReleaseKeyStateImmediatelyOnMainThreadAsync(
+            Keyboard keyboard,
+            Key key,
+            CancellationToken ct)
+        {
+            await InputSystemUpdateHelper.SwitchToMainThreadIfNeeded(ct);
+            ReleaseKeyStateImmediately(keyboard, key);
+        }
+
+        private static void ReleaseKeyStateImmediately(Keyboard keyboard, Key key)
+        {
+            Debug.Assert(CanInjectKeyboardState(keyboard), "keyboard state can only be released while PlayMode has a keyboard");
+            if (!CanInjectKeyboardState(keyboard))
+            {
+                return;
+            }
+
+            KeyboardKeyState.SetKeyState(keyboard, key, false);
+            InputSystemUpdateHelper.RunExplicitUpdate(InputUpdateTypeResolver.Resolve());
+        }
+
+        internal static void ScheduleTimedOutPressCleanup(Keyboard keyboard, Key key, bool pressWasApplied)
+        {
+            CleanupTimedOutPressAsync(keyboard, key, pressWasApplied, CancellationToken.None).Forget();
+        }
+
+        private static async Task CleanupTimedOutPressAsync(
+            Keyboard keyboard,
+            Key key,
+            bool pressWasApplied,
+            CancellationToken ct)
+        {
+            await InputSystemUpdateHelper.SwitchToMainThreadIfNeeded(ct);
+            if (pressWasApplied)
+            {
+                await ReleaseKeyStateIfPossible(keyboard, key, ct).ConfigureAwait(false);
+            }
+
+            KeyboardKeyState.UnregisterTransientKey(key);
+            SimulateKeyboardOverlayState.ClearPress();
+        }
+
+        internal static void ScheduleTimedOutHeldKeyCleanup(
+            Keyboard keyboard,
+            Key key,
+            string keyName,
+            bool keyWasApplied)
+        {
+            CleanupTimedOutHeldKeyAsync(keyboard, key, keyName, keyWasApplied, CancellationToken.None).Forget();
+        }
+
+        private static async Task CleanupTimedOutHeldKeyAsync(
+            Keyboard keyboard,
+            Key key,
+            string keyName,
+            bool keyWasApplied,
+            CancellationToken ct)
+        {
+            await InputSystemUpdateHelper.SwitchToMainThreadIfNeeded(ct);
+            if (keyWasApplied)
+            {
+                await ReleaseKeyStateIfPossible(keyboard, key, ct).ConfigureAwait(false);
+            }
+
+            KeyboardKeyState.SetKeyUp(key);
+            SimulateKeyboardOverlayState.RemoveHeldKey(keyName);
+        }
+
+        private static bool CanInjectKeyboardState(Keyboard keyboard)
+        {
+            return EditorApplication.isPlaying && keyboard != null;
+        }
+    }
+}
+#endif
