+++ Packages/src/Editor/FirstPartyTools/SimulateKeyboard/SimulateKeyboardUseCase.cs
@@ -2,8 +2,6 @@
 using System;
 using System.Threading;
 using System.Threading.Tasks;
-using UnityEditor;
-using UnityEngine;
 #if ULOOP_HAS_INPUT_SYSTEM
 using UnityEngine.InputSystem;
 #endif
@@ -205,16 +203,22 @@ namespace io.github.hatayama.UnityCliLoop.FirstPartyTools
                 InputSystem.onAfterUpdate -= pressEdgeMonitor;
                 if (waitOutcome == InputSimulationWaitOutcome.TimedOut)
                 {
-                    ScheduleTimedOutPressCleanup(keyboard, key, pressWasApplied);
+                    KeyboardInputMainThreadCleanup.ScheduleTimedOutPressCleanup(
+                        keyboard,
+                        key,
+                        pressWasApplied);
                 }
                 else if (pressWasApplied)
                 {
                     InputSimulationWaitOutcome releaseOutcome =
-                        await ReleaseKeyStateIfPossible(keyboard, key, CancellationToken.None).ConfigureAwait(false);
+                        await KeyboardInputMainThreadCleanup.ReleaseKeyStateIfPossible(
+                            keyboard,
+                            key,
+                            CancellationToken.None).ConfigureAwait(false);
                     if (releaseOutcome == InputSimulationWaitOutcome.TimedOut)
                     {
                         waitOutcome = InputSimulationWaitOutcome.TimedOut;
-                        ScheduleTimedOutPressCleanup(keyboard, key, false);
+                        KeyboardInputMainThreadCleanup.ScheduleTimedOutPressCleanup(keyboard, key, false);
                     }
                     else if (waitOutcome == InputSimulationWaitOutcome.Paused)
                     {
@@ -224,7 +228,7 @@ namespace io.github.hatayama.UnityCliLoop.FirstPartyTools
                     else
                     {
                         KeyboardKeyState.UnregisterTransientKey(key);
-                        await FinalizePressOverlay(ct).ConfigureAwait(false);
+                        await KeyboardInputMainThreadCleanup.FinalizePressOverlay(ct).ConfigureAwait(false);
                     }
                 }
                 else
@@ -309,12 +313,20 @@ namespace io.github.hatayama.UnityCliLoop.FirstPartyTools
                 InputSystem.onAfterUpdate -= keyDownEdgeMonitor;
                 if (waitOutcome == InputSimulationWaitOutcome.TimedOut)
                 {
-                    ScheduleTimedOutHeldKeyCleanup(keyboard, key, keyName, keyDownApplied);
+                    KeyboardInputMainThreadCleanup.ScheduleTimedOutHeldKeyCleanup(
+                        keyboard,
+                        key,
+                        keyName,
+                        keyDownApplied);
                 }
                 else if (keyDownApplied && !committed)
                 {
                     InputSimulationWaitOutcome rollbackOutcome =
-                        await RollbackHeldKey(keyboard, key, keyName, CancellationToken.None).ConfigureAwait(false);
+                        await KeyboardInputMainThreadCleanup.RollbackHeldKey(
+                            keyboard,
+                            key,
+                            keyName,
+                            CancellationToken.None).ConfigureAwait(false);
                     if (rollbackOutcome == InputSimulationWaitOutcome.TimedOut)
                     {
                         waitOutcome = InputSimulationWaitOutcome.TimedOut;
@@ -366,11 +378,14 @@ namespace io.github.hatayama.UnityCliLoop.FirstPartyTools
             }
 
             InputSimulationWaitOutcome releaseOutcome =
-                await ReleaseKeyStateIfPossible(keyboard, key, CancellationToken.None).ConfigureAwait(false);
+                await KeyboardInputMainThreadCleanup.ReleaseKeyStateIfPossible(
+                    keyboard,
+                    key,
+                    CancellationToken.None).ConfigureAwait(false);
 
             if (releaseOutcome == InputSimulationWaitOutcome.TimedOut)
             {
-                ScheduleTimedOutHeldKeyCleanup(keyboard, key, keyName, false);
+                KeyboardInputMainThreadCleanup.ScheduleTimedOutHeldKeyCleanup(keyboard, key, keyName, false);
                 return KeyboardInputSimulationResponseFactory.TimedOutKeyResult(
                     UnityCliLoopKeyboardAction.KeyUp,
                     keyName);
@@ -413,145 +428,6 @@ namespace io.github.hatayama.UnityCliLoop.FirstPartyTools
                 return Key.Enter.ToString();
             }
             return keyName;
-        }
-
-        private static async Task FinalizePressOverlay(CancellationToken ct)
-        {
-            await InputSystemUpdateHelper.SwitchToMainThreadIfNeeded(CancellationToken.None);
-            if (ct.IsCancellationRequested)
-            {
-                SimulateKeyboardOverlayState.ClearPress();
-                return;
-            }
-
-            SimulateKeyboardOverlayState.ReleasePress();
-            await EditorFrameWaiter.WaitFramesOrTimeoutAsync(
-                1,
-                UnityCliLoopConstants.EDITOR_FRAME_WAIT_TIMEOUT_MS,
-                CancellationToken.None).ConfigureAwait(false);
-        }
-
-        private static async Task<InputSimulationWaitOutcome> RollbackHeldKey(
-            Keyboard keyboard,
-            Key key,
-            string keyName,
-            CancellationToken ct)
-        {
-            await InputSystemUpdateHelper.SwitchToMainThreadIfNeeded(ct);
-            InputSimulationWaitOutcome releaseOutcome =
-                await ReleaseKeyStateIfPossible(keyboard, key, ct).ConfigureAwait(false);
-            if (releaseOutcome == InputSimulationWaitOutcome.TimedOut)
-            {
-                ScheduleTimedOutHeldKeyCleanup(keyboard, key, keyName, false);
-                return releaseOutcome;
-            }
-
-            await InputSystemUpdateHelper.SwitchToMainThreadIfNeeded(ct);
-            KeyboardKeyState.SetKeyUp(key);
-            SimulateKeyboardOverlayState.RemoveHeldKey(keyName);
-            return releaseOutcome;
-        }
-
-        private static async Task<InputSimulationWaitOutcome> ReleaseKeyStateIfPossible(
-            Keyboard keyboard,
-            Key key,
-            CancellationToken ct)
-        {
-            await InputSystemUpdateHelper.SwitchToMainThreadIfNeeded(ct);
-            if (!CanInjectKeyboardState(keyboard))
-            {
-                return InputSimulationWaitOutcome.Completed;
-            }
-
-            if (EditorApplication.isPaused)
-            {
-                ReleaseKeyStateImmediately(keyboard, key);
-                return InputSimulationWaitOutcome.Completed;
-            }
-
-            InputSimulationWaitOutcome releaseOutcome = await InputSystemUpdateHelper.ApplyOnNextConfiguredUpdate(
-                () => KeyboardKeyState.SetKeyState(keyboard, key, false),
-                ct).ConfigureAwait(false);
-            if (releaseOutcome == InputSimulationWaitOutcome.TimedOut)
-            {
-                ScheduleReleaseKeyStateImmediately(keyboard, key);
-            }
-
-            return releaseOutcome;
-        }
-
-        private static void ScheduleReleaseKeyStateImmediately(Keyboard keyboard, Key key)
-        {
-            ReleaseKeyStateImmediatelyOnMainThreadAsync(keyboard, key, CancellationToken.None).Forget();
-        }
-
-        private static async Task ReleaseKeyStateImmediatelyOnMainThreadAsync(
-            Keyboard keyboard,
-            Key key,
-            CancellationToken ct)
-        {
-            await InputSystemUpdateHelper.SwitchToMainThreadIfNeeded(ct);
-            ReleaseKeyStateImmediately(keyboard, key);
-        }
-
-        private static void ReleaseKeyStateImmediately(Keyboard keyboard, Key key)
-        {
-            Debug.Assert(CanInjectKeyboardState(keyboard), "keyboard state can only be released while PlayMode has a keyboard");
-            if (!CanInjectKeyboardState(keyboard))
-            {
-                return;
-            }
-
-            KeyboardKeyState.SetKeyState(keyboard, key, false);
-            InputSystemUpdateHelper.RunExplicitUpdate(InputUpdateTypeResolver.Resolve());
-        }
-
-        private static void ScheduleTimedOutPressCleanup(Keyboard keyboard, Key key, bool pressWasApplied)
-        {
-            CleanupTimedOutPressAsync(keyboard, key, pressWasApplied, CancellationToken.None).Forget();
-        }
-
-        private static async Task CleanupTimedOutPressAsync(
-            Keyboard keyboard,
-            Key key,
-            bool pressWasApplied,
-            CancellationToken ct)
-        {
-            await InputSystemUpdateHelper.SwitchToMainThreadIfNeeded(ct);
-            if (pressWasApplied)
-            {
-                await ReleaseKeyStateIfPossible(keyboard, key, ct).ConfigureAwait(false);
-            }
-
-            KeyboardKeyState.UnregisterTransientKey(key);
-            SimulateKeyboardOverlayState.ClearPress();
-        }
-
-        private static void ScheduleTimedOutHeldKeyCleanup(Keyboard keyboard, Key key, string keyName, bool keyWasApplied)
-        {
-            CleanupTimedOutHeldKeyAsync(keyboard, key, keyName, keyWasApplied, CancellationToken.None).Forget();
-        }
-
-        private static async Task CleanupTimedOutHeldKeyAsync(
-            Keyboard keyboard,
-            Key key,
-            string keyName,
-            bool keyWasApplied,
-            CancellationToken ct)
-        {
-            await InputSystemUpdateHelper.SwitchToMainThreadIfNeeded(ct);
-            if (keyWasApplied)
-            {
-                await ReleaseKeyStateIfPossible(keyboard, key, ct).ConfigureAwait(false);
-            }
-
-            KeyboardKeyState.SetKeyUp(key);
-            SimulateKeyboardOverlayState.RemoveHeldKey(keyName);
-        }
-
-        private static bool CanInjectKeyboardState(Keyboard keyboard)
-        {
-            return EditorApplication.isPlaying && keyboard != null;
         }
 
         // Runs inside InputSystem.onAfterUpdate. Editor updates are excluded because a press
